modified:   element_smapshot.js 	deleted:    element_smapshot_v2.js 	new file:   map_smapshot.js
